requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
fastmcp>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
orjson>=3.10.0
